+++ packages/vuikit/src/table/elements/index.js
@@ -2,6 +2,5 @@ export { default as ElTable } from './table'
 export { default as ElTableTd } from './table-td'
 export { default as ElTableTr } from './table-tr'
 export { default as ElTableTh } from './table-th'
-export { default as ElTableThSort } from './table-th-sort'
 
 export { default as ElTableCellCheckbox } from './table-cell-checkbox'